Add base classes for material-related classes. Pass pointers to NimbleMain
