+++ src/nimble_block.h
@@ -44,8 +44,9 @@
 #ifndef NIMBLE_BLOCK_H
 #define NIMBLE_BLOCK_H
 
+#include "nimble_block_base.h"
+
 #include "nimble_element.h"
-#include "nimble_material.h"
 #include "nimble_genesis_mesh.h"
 
 #ifdef NIMBLE_HAVE_DARMA
@@ -62,11 +63,11 @@ namespace nimble {
 
   class UqModel;//Forward declaration to avoid circular header inclusion
 
-  class Block {
+  class Block : public nimble::BlockBase {
 
   public:
 
-    Block() :
+    Block() : BlockBase(),
       macro_material_parameters_("none"), 
       vol_ave_volume_offset_(-1), 
       rve_boundary_condition_strategy_("none") 
@@ -76,7 +77,7 @@ namespace nimble {
 #endif
       {}
 
-    virtual ~Block() { }
+    ~Block() = default;
 
 #ifdef NIMBLE_HAVE_DARMA
     template<typename ArchiveType>
@@ -124,21 +125,9 @@ namespace nimble {
 
     void InstantiateMaterialModel(MaterialFactory& factory);
 
-    void InstantiateElement();
+    void InstantiateElement() override;
 
     void GetDataLabelsAndLengths(std::vector< std::pair<std::string, Length> >& data_labels_and_lengths);
-
-    double GetDensity() const {
-      return material_->GetDensity();
-    }
-
-    double GetBulkModulus() const {
-      return material_->GetBulkModulus();
-    }
-
-    double GetShearModulus() const {
-      return material_->GetShearModulus();
-    }
 
     void ComputeLumpedMassMatrix(const double * const reference_coordinates,
                                  int num_elem,
@@ -199,9 +188,9 @@ namespace nimble {
                                    int num_derived_elem_data,
                                    std::vector< std::vector<double> >& derived_elem_data);
 
-    std::shared_ptr<Material> const GetMaterialPointer() const { return material_; }
+    std::shared_ptr<Material> GetMaterialPointer() const { return material_; }
 
-    std::shared_ptr<Element> const GetElementPointer() const { return element_; }
+    std::shared_ptr<Element> GetElementPointer() const { return element_; }
 
 #ifdef NIMBLE_HAVE_UQ
     // HACK specific to elastic models
@@ -215,7 +204,7 @@ namespace nimble {
     }
 #endif
 
-  private:
+  protected:
 
     void DetermineDataOffsets(std::vector<std::string> const & elem_data_labels,
                               std::vector<std::string> const & derived_elem_data_labels);
@@ -226,8 +215,6 @@ namespace nimble {
     std::vector<int> rve_output_global_elem_ids_;
     // todo: can we avoid carrying the rve_mesh around?
     GenesisMesh rve_mesh_;
-    std::shared_ptr<Element> element_ = nullptr;
-    std::shared_ptr<Material> material_ = nullptr;
     std::vector<int> def_grad_offset_;
     std::vector<int> stress_offset_;
     std::vector<int> state_data_offset_;
