+++ src/nimble_data_manager.h
@@ -49,6 +49,7 @@
 #include "nimble_exodus_output.h"
 #include "nimble_genesis_mesh.h"
 #include "nimble_linear_solver.h"
+#include "nimble_model_data.h"
 
 #ifdef NIMBLE_HAVE_DARMA
   #include "darma.h"
@@ -59,168 +60,6 @@
 #endif
 
 namespace nimble {
-
-  class BaseModelData {
-
-  public:
-
-    BaseModelData() = default;
-
-    virtual ~BaseModelData() = default;
-
-    void SetDimension(int dim);
-
-    int GetDimension() const {return dim_; }
-
-    void SetCriticalTimeStep(double time_step) { critical_time_step_ = time_step; }
-
-    double GetCriticalTimeStep() const { return critical_time_step_; }
-
-    const std::vector<std::string> & GetNodeDataLabelsForOutput() const {
-      return output_node_component_labels_;
-    }
-
-    const std::map<int, std::vector<std::string> > & GetElementDataLabels() const {
-      return element_component_labels_;
-    }
-
-    const std::map<int, std::vector<std::string> > & GetElementDataLabelsForOutput() const {
-      return output_element_component_labels_;
-    }
-
-    const std::map<int, std::vector<std::string> > & GetDerivedElementDataLabelsForOutput() const {
-      return derived_output_element_data_labels_;
-    }
-
-    void SetDerivedElementDataLabelsForOutput(std::map<int, std::vector<std::string> > &&ref)
-    {
-      derived_output_element_data_labels_ = ref;
-    }
-
-    virtual int AllocateNodeData(Length length,
-                         std::string label,
-                         int num_objects) = 0;
-
-    virtual int GetFieldId(const std::string& field_label) const = 0;
-
-    virtual void SetReferenceCoordinates(const nimble::GenesisMesh &mesh) = 0;
-
-  protected:
-
-    //! Problem dimension, either 2 or 3.
-    int dim_ = 3;
-
-    //! Critical time step
-    double critical_time_step_ = 0.0;
-
-    //! Output labels for node data that will be written to disk
-    std::vector<std::string> output_node_component_labels_;
-
-    //! Map key is the block_id, vector contains component-wise label for each scalar entry in the data array.
-    std::map<int, std::vector<std::string>> element_component_labels_;
-
-    //! Output labels for element data that will be written to disk.
-    std::map<int, std::vector<std::string> > output_element_component_labels_;
-
-    //! Output labels for derived element data that will be written to disk.
-    std::map<int, std::vector<std::string> > derived_output_element_data_labels_;
-
-  };
-
-  class ModelData : public BaseModelData {
-
-  public:
-
-    ModelData() = default;
-
-    ~ModelData() override = default;
-
-#ifdef NIMBLE_HAVE_DARMA
-    template<typename ArchiveType>
-    void serialize(ArchiveType& ar) {
-      ar | dim_ | critical_time_step_ | block_ids_ | blocks_ | data_fields_ | node_data_ | output_node_component_labels_ | element_data_fields_ | element_component_labels_ | element_data_n_ | element_data_np1_ | output_element_component_labels_ | derived_output_element_data_labels_ | globally_shared_nodes_ | global_node_id_to_local_node_id_;
-    }
-#endif
-
-    int GetFieldId(const std::string& label) const override;
-
-    Field GetField(int field_id);
-
-    int AllocateNodeData(Length length,
-                         std::string label,
-                         int num_objects) override;
-
-    double* GetNodeData(int field_id);
-
-    void GetNodeDataForOutput(std::vector< std::vector<double> >& single_component_arrays);
-
-    void DeclareElementData(int block_id,
-                            std::vector< std::pair<std::string, Length> > const & data_labels_and_lengths);
-
-    void AllocateElementData(std::map<int, int> const & num_integration_points_in_each_block);
-
-    std::vector<double> & GetElementDataOld(int block_id) {
-      return element_data_n_.at(block_id);
-    }
-
-    std::vector<double> & GetElementDataNew(int block_id) {
-      return element_data_np1_.at(block_id);
-    }
-
-    void GetElementDataForOutput(std::map<int, std::vector< std::vector<double> > >& single_component_arrays);
-
-    void SpecifyOutputFields(std::string output_field_string);
-
-    std::map<int, Block>& GetBlocks() { return blocks_; }
-
-    std::map<int, Block> const & GetBlocks() const { return blocks_; }
-
-    std::vector<int>& GetGloballySharedNodes() { return globally_shared_nodes_; }
-
-    std::map<int, int>& GetGlobalNodeIdToLocalNodeIdMap() { return global_node_id_to_local_node_id_; }
-
-    void SwapStates() {
-      element_data_n_.swap(element_data_np1_);
-    }
-
-   void SetReferenceCoordinates(const nimble::GenesisMesh &mesh) override;
-
-  protected:
-
-    void AssignFieldId(Field& field);
-
-    void GetNodeDataComponent(int field_id,
-                              int component,
-                              double* const component_data);
-
-    //! Block ids
-    std::vector<int> block_ids_;
-
-    //! Blocks
-    std::map<int, Block> blocks_;
-
-    //! Map key is the field_id, value is the corresponding Field.
-    std::map<int, Field> data_fields_;
-
-    //! Map key is the field_id, vector contains the nested data array for the given nodal field.
-    std::map<int, std::vector<double>> node_data_;
-
-    //! Map key is the block_id, the vector contains the field ids for the fields on that block.
-    std::map<int, std::vector<int>> element_data_fields_;
-
-    //! Map key is the block_id, vector contains full data array for that block at step N.
-    std::map<int, std::vector<double> > element_data_n_;
-
-    //! Map key is the block_id, vector contains full data array for that block at step N+1.
-    std::map<int, std::vector<double> > element_data_np1_;
-
-    //! List of node ids that are shared across multiple ranks
-    std::vector<int> globally_shared_nodes_;
-
-    //! Map from global node it to local node id
-    std::map<int, int> global_node_id_to_local_node_id_;
-
-  };
 
   struct RVEData {
     ModelData model_data_;
@@ -243,9 +82,9 @@ namespace nimble {
 
   public:
 
-    DataManager() {}
+    DataManager() = default;
 
-    virtual ~DataManager() {}
+    virtual ~DataManager() = default;
 
 #ifdef NIMBLE_HAVE_DARMA
     template<typename ArchiveType>
