+++ src/nimble_kokkos_block_material_interface.cc
@@ -1,11 +1,11 @@
-#include <nimble_kokkos_block_material_interface.h>
-#include <nimble_material.h>
+#include "nimble_kokkos_block_material_interface.h"
+#include "nimble_material.h"
 
 namespace nimble_kokkos {
 
 namespace {
 
-inline void compute_block_stress(const BlockData& block_data,
+inline void compute_block_stress(const nimble::BlockData& block_data,
                                  nimble_kokkos::DeviceFullTensorIntPtView deformation_gradient_step_n_d,
                                  nimble_kokkos::DeviceFullTensorIntPtView deformation_gradient_step_np1_d,
                                  nimble_kokkos::DeviceSymTensorIntPtView stress_step_n_d,
@@ -34,13 +34,13 @@ inline void compute_block_stress(const BlockData& block_data,
 
 void BlockMaterialInterface::ComputeStress() const {
   for (auto &&block_data : blocks) {
-    auto deformation_gradient_step_n_d = model_data.GetDeviceFullTensorIntegrationPointData(
+    auto deformation_gradient_step_n_d = model_data->GetDeviceFullTensorIntegrationPointData(
         block_data.id, field_ids.deformation_gradient, nimble::STEP_N);
-    auto deformation_gradient_step_np1_d = model_data.GetDeviceFullTensorIntegrationPointData(
+    auto deformation_gradient_step_np1_d = model_data->GetDeviceFullTensorIntegrationPointData(
         block_data.id, field_ids.deformation_gradient, nimble::STEP_NP1);
-    auto stress_step_n_d = model_data.GetDeviceSymTensorIntegrationPointData(block_data.id, field_ids.stress,
+    auto stress_step_n_d = model_data->GetDeviceSymTensorIntegrationPointData(block_data.id, field_ids.stress,
                                                                              nimble::STEP_N);
-    auto stress_step_np1_d = model_data.GetDeviceSymTensorIntegrationPointData(block_data.id, field_ids.stress,
+    auto stress_step_np1_d = model_data->GetDeviceSymTensorIntegrationPointData(block_data.id, field_ids.stress,
                                                                                nimble::STEP_NP1);
     compute_block_stress(block_data, deformation_gradient_step_n_d, deformation_gradient_step_np1_d,
                          stress_step_n_d, stress_step_np1_d, time_n, time_np1);
