+++ src/nimble_kokkos_block_material_interface_factory.cc
@@ -41,45 +41,24 @@
 //@HEADER
 */
 
-#include <stdexcept>
-#include <tuple>
-#include <utility>
-#include <nimble_kokkos_defs.h>
-#include <nimble_kokkos_material_factory.h>
-#include <nimble_material.h>
+#include <memory>
+
+#include "nimble_kokkos_block_material_interface_factory.h"
+#include "nimble_kokkos_block_material_interface.h"
+
 
 namespace nimble_kokkos {
 
-using nimble::Material;
-
-MaterialFactory::MaterialFactory()
-    : MaterialFactoryBase(),
-      material_device(nullptr) {
-}
-
-template <typename MatType>
-inline std::pair<std::shared_ptr<MatType>, MatType*> allocate_material_on_host_and_device(const nimble::MaterialParameters& mat_params_struct) {
-  auto mat_host = std::make_shared<MatType>(mat_params_struct);
-  auto mat_device = static_cast<MatType*>(Kokkos::kokkos_malloc<>("Material", sizeof(MatType)));
-  MatType& mat_host_ref = *mat_host;
-  Kokkos::parallel_for(1, KOKKOS_LAMBDA(int) {
-    new (mat_device) MatType(mat_host_ref);
-  });
-  Kokkos::fence();
-  return std::make_pair(mat_host, mat_device);
-}
-
-void MaterialFactory::create() {
-  auto name_string = material_params->GetMaterialName(false);
-  if (name_string == "neohookean") {
-    std::tie(material, material_device) = allocate_material_on_host_and_device<nimble::NeohookeanMaterial>(
-        *material_params);
-  } else if (name_string == "elastic") {
-    std::tie(material, material_device) = allocate_material_on_host_and_device<nimble::ElasticMaterial>(
-        *material_params);
-  } else {
-    throw std::logic_error("\nError in Block::InstantiateMaterialModel(), invalid material model name.\n");
-  }
+std::shared_ptr<nimble::BlockMaterialInterfaceBase> BlockMaterialInterfaceFactory::create
+(
+    double time_n, double time_np1,
+    const nimble::FieldIds &field_ids,
+    const std::vector<nimble::BlockData> &blocks,
+    nimble::BaseModelData *model_data_ptr
+) const
+{
+  std::shared_ptr<nimble::BlockMaterialInterfaceBase> res_ptr(new nimble_kokkos::BlockMaterialInterface(time_n, time_np1, field_ids, blocks, model_data_ptr));
+  return res_ptr;
 }
 
 }
