+++ src/nimble_mpi_main.cc
@@ -43,6 +43,7 @@
 
 #include <memory>
 
+#include "nimble_block_material_interface_factory_base.h"
 #include "nimble_contact_interface.h"
 #include "nimble_material_factory.h"
 #include "nimble_mpi.h"
@@ -53,9 +54,11 @@ int main(int argc, char *argv[]) {
   nimble::Parser myParser;
   nimble::NimbleInitializeAndGetInput(argc, argv, myParser);
 
+  std::shared_ptr<MaterialFactoryType> material_factory(new nimble::MaterialFactory);
+  std::shared_ptr<nimble::BlockMaterialInterfaceFactoryBase> block_material = nullptr;
   std::shared_ptr<nimble::ContactInterface> contact_interface(new nimble::ContactInterface);
-  std::shared_ptr<MaterialFactoryType> material_factory(new MaterialFactoryType);
-  int status = nimble::NimbleMain(material_factory, contact_interface, myParser);
+
+  int status = nimble::NimbleMain(material_factory, contact_interface, block_material, myParser);
 
   nimble::NimbleFinalize(myParser);
